backend/requirements.txt
# -*- coding: utf-8 -*-
# 工程预装模块
# 需要额外的python包，可直接在文件后面添加
# 请确保指定的包和版本号，可通过pip安装

# blueapps requirement

redis==3.5.3
Django==3.2.4
celery==4.4.0
eventlet==0.31.0
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-cors-headers==3.7.0

# jwt
pyCryptodome==3.9.7
PyJWT==1.7.1
cryptography==2.7
mysqlclient==2.0.3
MarkupSafe==1.1.1
Mako==1.1.4
requests==2.25.1
python-json-logger==2.0.1
whitenoise==5.2.0
gunicorn==19.6.0

blueapps==4.1.1

djangorestframework==3.12.4
django-filter==2.4.0